backend/requirements.txt
fastapi==0.78.0
pandas==1.3.4
Pillow==9.1.1
torch==1.10.0
torchvision==0.11.1
uvicorn==0.17.6
python-multipart
